+++ fabric/src/main/java/de/maxhenkel/voicechat/FabricVoicechatMod.java
@@ -4,6 +4,8 @@ import de.maxhenkel.configbuilder.ConfigBuilder;
 import de.maxhenkel.voicechat.config.FabricServerConfig;
 import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
 import de.maxhenkel.voicechat.intercompatibility.FabricCommonCompatibilityManager;
+import de.maxhenkel.voicechat.permission.FabricPermissionManager;
+import de.maxhenkel.voicechat.permission.PermissionManager;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.loader.api.FabricLoader;
@@ -32,6 +34,11 @@ public class FabricVoicechatMod extends Voicechat implements ModInitializer {
     @Override
     public CommonCompatibilityManager createCompatibilityManager() {
         return new FabricCommonCompatibilityManager();
+    }
+
+    @Override
+    protected PermissionManager createPermissionManager() {
+        return new FabricPermissionManager();
     }
 
 }